pyproject.toml
[tool.poetry]
name = "test"
version = "0.1.0"
description = "UI & API tests"
authors = ["OlesiaChirkova <chirkova.o@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^8.1.1"
requests = "^2.31.0"
faker = "^24.3.0"
python-dotenv = "^1.0.1"
allure-pytest = "^2.13.3"
pydantic = "^2.6.4"
playwright = "^1.42.0"
config = "^0.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
